add search for aiports, create flight and add fixes
