+++ front/src/views/Manager/NextFlightListing/components/RegisterFlightDialog.vue
@@ -16,14 +16,14 @@ import SelectGroup from '@/components/ui/select/SelectGroup.vue'
 import SelectItem from '@/components/ui/select/SelectItem.vue'
 import SelectTrigger from '@/components/ui/select/SelectTrigger.vue'
 import SelectValue from '@/components/ui/select/SelectValue.vue'
-import { loadAllAirpoirts } from '@/mock/airpoirts'
+import { loadAllAirpoirts } from '../services/NextFlightListingService'
 import { useGlobalStore } from '@/stores/global'
 import type { Airport } from '@/types/Airpoirt'
 import { computed, onMounted, ref } from 'vue'
 import { useRegisterFlightForm } from '../composables/useRegisterFlightForm'
 import { extractPriceValue } from '@/utils/currency/extractPriceValue'
 import { currencyMask } from '@/utils/currency/inputMask'
-import { registerFlight } from '@/mock/flight'
+import { registerFlight } from '../services/NextFlightListingService'
 import { generateFlightCode } from '@/utils/generateFlightCode'
 
 defineProps<{
@@ -76,10 +76,13 @@ const onSubmit = handleSubmit(async (data) => {
 
     const generatedCode = generateFlightCode()
     await registerFlight({
-      ...data,
-      code: generatedCode,
-      price: price.value,
-    })
+      data: new Date(data.date).toISOString(),
+      valor_passagem: price.value,
+      quantidade_poltronas_total: data.seatsNumber,
+      quantidade_poltronas_ocupadas: 0,
+      codigo_aeroporto_origem: data.originAirport,
+      codigo_aeroporto_destino: data.destinationAirport
+})
 
     globalStore.setNotification({
       title: 'Voo cadastrado com sucesso',
